+++ src/main/java/com/gcu/realestate/Business/HousesServiceInterface.java
@@ -8,7 +8,7 @@ public interface HousesServiceInterface {
     
     public void test();
     public List<ProductModel> getHouses();
-    public ProductModel getOne(int id);
+    public ProductModel getOne(Long id);
     public List<ProductModel> searchProduct(String searchTerm);
     public int addOne(ProductModel newProduct);
     public boolean deleteOne(long l);
